feat(CoreModule): Added CoreModuleType interface to enable the creation of custom modules
